mots-data-ingress-master/twitter/requirements.txt
python-dotenv
peewee==3.14.0
pymysql
elasticsearch
tweepy
